Made the DataAssociation class sealed.
In translator updated the type declaration of translator to be an optional type. Adjusted all other method calls tp the object ignore the poss of a null call as this is checked before hand by IsTranslatorOperational boolean attribute.

Changed TranslationService.cs to being a file-scoped namespace from a block-scoped namespace.